Document expansion & fmt

+++ src/lib/builtins/helpers.rs
@@ -35,7 +35,7 @@ impl Status {
 
     pub fn is_failure(self) -> bool { self.0 != 0 }
 
-    pub fn as_os_code(self) -> i32 { self.0 as i32 }
+    pub fn as_os_code(self) -> i32 { self.0 }
 
     pub fn toggle(&mut self) { self.0 = if self.is_success() { 1 } else { 0 }; }
 }
